Add files via upload
We asked students to submit URLs of online certificates they obtain on an online learning platform. This piece of code demonstrates getting students URLs submissions from Canvas, verifying the URLs, scrapping the content of the webpage and checking if the name of the course and the word "Verified" are on the website.